Refactor code

+++ backend/index.js
@@ -1,9 +1,11 @@
-const express = require("express");
 const bodyParser = require("body-parser");
-const { Pool } = require("pg");
 const cors = require("cors"); // Import the cors middleware
 const config = require("./config.js");
+const express = require("express");
 const jwt = require("jsonwebtoken");
+const { Pool } = require("pg");
+
+const TOKEN_EXPIRE_TIME = 60000; // in milliseconds
 
 const app = express();
 app.use(bodyParser.json());
@@ -25,43 +27,29 @@ const pool = new Pool({
 
 function verifyToken(req, res, next) {
   const authHeader = req.header("Authorization");
+
   if (!authHeader) {
     return res.status(401).json({ error: `Access denied. No token provided.` });
   }
+
   if (!authHeader?.startsWith("Bearer ")) {
     return res.status(401).json({ error: `Access denied. Not Bearer Token.` });
   }
+
   const token = authHeader.substring(7, authHeader.length);
+
   try {
     const decoded = jwt.verify(token, config.jwtSecret);
     const currentTimeInSeconds = Date.now();
     if (decoded.exp && currentTimeInSeconds > decoded.exp) {
       return res.status(401).json({ error: "Token has expired." });
     }
-    req.user = decoded.username; // Store the decoded user information in the request object
+    req.user = decoded.username;
     next(); // Move to the next middleware
   } catch (error) {
     res.status(400).json({ error: `Error while verifying token. ${error}` });
   }
 }
-
-// GET /api/users/:username - Retrieve user details by username
-app.get("/api/users/:username", verifyToken, async (req, res) => {
-  const username = req.params.username;
-  try {
-    const query =
-      "SELECT user_id, username, email, firstname, lastname FROM userAccounts WHERE username = $1";
-    const result = await pool.query(query, [username]);
-    if (result.rows.length === 0) {
-      res.status(404).json({ error: "User not found" });
-    } else {
-      res.json(result.rows[0]);
-    }
-  } catch (error) {
-    console.error("Error retrieving user:", error);
-    res.status(500).json({ error: "Fail to receive username." });
-  }
-});
 
 // POST /api/login - Login with JWT
 app.post("/api/login", async (req, res) => {
@@ -73,9 +61,8 @@ app.post("/api/login", async (req, res) => {
     if (result.rows.length === 0) {
       res.status(401).json({ error: "Incorrect username or Password" });
     } else {
-      // Create a JWT token
       const user = result.rows[0];
-      const exp = Date.now() + 1000 * 60 * 1;
+      const exp = Date.now() + TOKEN_EXPIRE_TIME;
       const token = jwt.sign(
         { username: user.username, role: user.role, exp: exp },
         config.jwtSecret
@@ -100,7 +87,7 @@ app.get("/api/refresh", verifyToken, async (req, res) => {
     const token = authHeader.substring(7, authHeader.length);
     const decoded = jwt.verify(token, config.jwtSecret);
 
-    const newExp = Date.now() + 1000 * 60 * 1;
+    const newExp = Date.now() + TOKEN_EXPIRE_TIME;
     const newToken = jwt.sign(
       { username: decoded.username, role: decoded.role, exp: newExp },
       config.jwtSecret
@@ -114,6 +101,24 @@ app.get("/api/refresh", verifyToken, async (req, res) => {
     });
   } catch (error) {
     res.status(400).json({ error: `Error Refreshing Token: ${error}` });
+  }
+});
+
+// GET /api/users/:username - Retrieve user details by username
+app.get("/api/users/:username", verifyToken, async (req, res) => {
+  const username = req.params.username;
+  try {
+    const query =
+      "SELECT user_id, username, email, firstname, lastname FROM userAccounts WHERE username = $1";
+    const result = await pool.query(query, [username]);
+    if (result.rows.length === 0) {
+      res.status(404).json({ error: "User not found" });
+    } else {
+      res.json(result.rows[0]);
+    }
+  } catch (error) {
+    console.error("Error retrieving user:", error);
+    res.status(500).json({ error: "Fail to receive username." });
   }
 });
 
@@ -141,14 +146,17 @@ app.post("/api/users", async (req, res) => {
 
 // PUT /api/users/ - Update user details
 app.put("/api/users", verifyToken, async (req, res) => {
-  const { oldUsername, newUsername, email, password, firstname, lastname } =
-    req.body;
+  const { username, email, password, firstname, lastname } = req.body;
+
+  if (req.user !== username) {
+    return res.status(401).json({ error: "Access denied." });
+  }
+
   try {
     const query =
-      "UPDATE userAccounts SET username = $2, email = $3, firstname = $4, lastname = $5 WHERE username = $1 RETURNING *";
+      "UPDATE userAccounts SET email = $2, firstname = $3, lastname = $4 WHERE username = $1 RETURNING *";
     const result = await pool.query(query, [
-      oldUsername,
-      newUsername,
+      username,
       email,
       firstname,
       lastname,
@@ -167,6 +175,11 @@ app.put("/api/users", verifyToken, async (req, res) => {
 // DELETE /api/users/:username - Delete a user by username
 app.delete("/api/users/:username", verifyToken, async (req, res) => {
   const username = req.params.username;
+
+  if (req.user !== username) {
+    return res.status(401).json({ error: "Access denied." });
+  }
+
   try {
     const query = "DELETE FROM userAccounts WHERE username = $1 RETURNING *";
     const result = await pool.query(query, [username]);
